requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
Flask==0.12.2
Flask-Classy==0.6.10
Flask-Cors==3.0.4
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.6.1
six==1.11.0
Werkzeug==0.14.1
